EmbeddedResource: Obsolete partial class removed
